requirements.txt
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
lxml==4.6.3
Pillow==8.3.1
pkg_resources==0.0.0
pycurl==7.44.0
requests==2.26.0
soupsieve==2.2.1
urllib3==1.26.6
